% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp.sim-package.R
\docType{package}
\name{amp.sim-package}
\alias{amp.sim}
\alias{amp.sim-package}
\title{amp.sim: Flexible Simulation Utilities for Pharmacometric Modeling}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of 'amp.sim' is to transform 'NONMEM' models into R syntax so they can be used for simulations using the 'deSolve', 'nlmixr2' or 'mrgsolve' package. Additionally, functionality is included to aid simulations performed directly in 'NONMEM' and to automatically create shiny apps for simulation models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://leidenadvancedpkpd.github.io/amp.sim/}
  \item \url{https://github.com/LeidenAdvancedPKPD/amp.sim}
  \item Report bugs at \url{https://github.com/LeidenAdvancedPKPD/amp.sim/issues}
}

}
\author{
\strong{Maintainer}: Richard Hooijmaijers \email{richardhooijmaijers@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Richard Hooijmaijers \email{richardhooijmaijers@gmail.com} [copyright holder]
}

Other contributors:
\itemize{
  \item LAPP Consultants \email{info@lapp.nl} [funder, copyright holder]
}

}
\keyword{internal}
