% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iiv_ruv_ctl.r
\name{get_iiv_ruv_ctl}
\alias{get_iiv_ruv_ctl}
\title{Get omega or sigma values from control stream}
\usage{
get_iiv_ruv_ctl(mod, type = "OMEGA")
}
\arguments{
\item{mod}{character vector of length 1 with filename of model, in case length is greater than 1 it is assumed to be a vector with model code}

\item{type}{character vector of length 1 with the type of block to get (valid types are "OMEGA" and "SIGMA")}
}
\value{
a named vector with omega values
}
\description{
This function gets the omega or sigma values from a model control stream. A separate function is developed as
obtaining these values is more difficult compared using to output
}
\examples{

\dontrun{
  get_omega_ctl("run1.mod")
  get_omega_ctl("run1.mod","SIGMA")
  get_omega_ctl(c("$OMEGA BLOCK(3)","0.2","0.1  0.3","0.02 0.03 0.04"))
}
}
\author{
Richard Hooijmaijers
}
