% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nmblock.r
\name{get_nmblock}
\alias{get_nmblock}
\title{Get information from dollar blocks inside NONMEM control streams}
\usage{
get_nmblock(model, block, ret = "content", omitbn = TRUE)
}
\arguments{
\item{model}{character vector of length 1 with filename of model, in case length is greater than 1 it is assumed to be a vector with model code}

\item{block}{character vector with names of the model blocks. Take into account that grep is used with respect to partial matching}

\item{ret}{character with the type of return value can be either "content" or "index"}

\item{omitbn}{logical indicating if the name of the block should be omitted when return (has only effect if ret="content")}
}
\value{
a list with either a numeric vector with the indices or a character vector with the content of the dollar blocks
}
\description{
This function returns a list with indices or content of dollar blocks
}
\examples{

mod  <- system.file("example_models","PK.1CMT.ORAL.mod", package = "amp.sim")
get_nmblock(mod,"OMEGA")
}
\author{
Richard Hooijmaijers
}
