% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.r
\name{get_param}
\alias{get_param}
\title{Get parameter values from model or ext file}
\usage{
get_param(model, lstblock, ext = NULL, addparam = TRUE)
}
\arguments{
\item{model}{character vector with the model content}

\item{lstblock}{list with each item being a separate structured dollar block, usually obtain from \code{\link{nmlistblock}}}

\item{ext}{character with the name of the NONMEM ext file (if not provided estimates are read directly from the list block)}

\item{addparam}{logical indicating if the function should try to add parameters (besides THETAs and the ones defined in covariates)
the additional parameters are always returned so it can be used for warnings}
}
\value{
a list with parameters, names and matrices
}
\description{
This function gets the parameter values from a NONMEM model or ext file including naming and additional
variables if present
}
\examples{

mod    <- system.file("example_models","PK.1CMT.ORAL.COV.mod", package = "amp.sim")
mdll   <- get_nmblock(mod,block=c("PK","DES"))
mdlls  <- nmlistblock(mdll)
get_param(mod, mdlls)

}
\author{
Richard Hooijmaijers
}
