% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm2mrgsolve.r
\name{nm2mrgsolve}
\alias{nm2mrgsolve}
\title{Convert NONMEM model to mrgsolve syntax}
\usage{
nm2mrgsolve(lstblock, model, ext = NULL, mod_return = NULL, out = NULL)
}
\arguments{
\item{lstblock}{structured list with information of the model that was read-in, usually obtained from the \link{nmlistblock} function}

\item{model}{character vector with the model content}

\item{ext}{character with the name of the NONMEM ext file (if not provided estimates are read directly from control stream)}

\item{mod_return}{a character vector indicating which items should be returned from the model function (see \link{convert_nonmem} for more details)}

\item{out}{character with the name of the output file without a file extension}
}
\value{
a list is returned inluding all building blocks to create a model
}
\description{
This function converts a NONMEM model to syntax useable in mrgsolve simulations.
}
\examples{

\dontrun{
  mdl <- readLines("run1.mod")
  bl  <- c("PROB","SUB","MODEL","PK","DES","PRED","THETA","OMEGA","ERROR","SIGMA","EST")
  lst <- get_nmblock(mdl,block=bl)
  lst <- nmlistblock(lst)
  nm2mrgsolve(lst)
}
}
\author{
Richard Hooijmaijers
}
