% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm2nonmem2rx.r
\name{nm2nonmem2rx}
\alias{nm2nonmem2rx}
\title{Convert NONMEM model to nonmem2rx syntax}
\usage{
nm2nonmem2rx(mod, out = NULL, control = "")
}
\arguments{
\item{mod}{character with the model file to be read in and converted}

\item{out}{character with the name of the output file without a file extension}

\item{control}{character with the type of control to bre returned (see \code{\link{convert_nonmem}} for more details)}
}
\value{
a list is returned inluding all building blocks to create a model
}
\description{
This function converts a NONMEM model to syntax useable in rxode2 simulations fully based on the nonmem2rx package
}
\examples{

\dontrun{
  nm2nonmem2rx("run1.mod")
}
}
\author{
Richard Hooijmaijers
}
