% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm2rxode2.r
\name{nm2rxode2}
\alias{nm2rxode2}
\title{Convert NONMEM model to rxode2 syntax}
\usage{
nm2rxode2(lstblock, ext = NULL, out = NULL, control = "")
}
\arguments{
\item{lstblock}{structured list with information of the model that was read-in, usually obtained from the \code{\link{nmlistblock}} function}

\item{ext}{character with the name of the NONMEM ext file (if not provided estimates are read directly from control stream)}

\item{out}{character with the name of the output file without a file extension}

\item{control}{character with the type of control to bre returned (see \code{\link{convert_nonmem}} for more details)}
}
\value{
a list is returned inluding all building blocks to create a model
}
\description{
This function converts a NONMEM model to syntax useable in rxode2 simulations
}
\examples{

\dontrun{
  mdl <- readLines("run1.mod")
  lst <- get_nmblock(mdl,block=c("PROB","SUB","MODEL","PK","DES","PRED","THETA","OMEGA","ERROR","SIGMA","EST"))
  lst <- nmlistblock(lst)
  nm2rxode2(lst)
}
}
\author{
Richard Hooijmaijers
}
