% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{pos_clpar}
\alias{pos_clpar}
\title{Get position of closing parenthesis}
\usage{
pos_clpar(x)
}
\arguments{
\item{x}{character string for which the parenthesis should be searched}
}
\value{
a numeric with the position of the closing parenthesis (will be -1 if no match is present)
}
\description{
This function gets the position of the closing parenthesis after the first opening one
}
\examples{

\dontrun{
  tst <- "IF (test == A(1)) a(1)=(1*5)/2"
  pos_clpar(tst)
  substring(tst,1,pos_clpar(tst))
}
}
\author{
Richard Hooijmaijers
}
