% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_pow.r
\name{conv_pow}
\alias{conv_pow}
\title{Convert infix power notation to prefix notation}
\usage{
conv_pow(x)
}
\arguments{
\item{x}{character vector with the formulas to convert}
}
\value{
a vector with the transformed power notations
}
\description{
This function will transform the power notation that can be used in R to the one needed by mrgsolve
}
\examples{

conv_pow("y = par1*(par2/par3)^xy + a - par4**(2/par5)")
}
\author{
Richard Hooijmaijers
}
