% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_nmsyntax.r
\name{convert_nmsyntax}
\alias{convert_nmsyntax}
\title{Convert NONMEM specific syntax to R syntax}
\usage{
convert_nmsyntax(x, type = "mrgsolve")
}
\arguments{
\item{x}{character vector with the syntax to be converted}

\item{type}{character with the type of syntax to convert to (currently "deSolve" and "mrgsolve" and "RxODE" are supported)}
}
\value{
character vector with the converted syntax
}
\description{
This function converts a NONMEM syntax to R syntax, mainly for functions and operators
}
\examples{

convert_nmsyntax("IF(VAR.GT.0) VAR2 = PHI(1)")
}
\author{
Richard Hooijmaijers
}
