% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_func.r
\name{dose_func}
\alias{dose_func}
\title{Create data.frame with dosing for usage in simulation}
\usage{
dose_func(cmt, value, tinf, tau, ndose, times)
}
\arguments{
\item{cmt}{number of compartment or differential equation where the dosing should be given}

\item{value}{the value of the dosing that should be given}

\item{tinf}{in case value is set an infusion is assumed with tinf as infusion time}

\item{tau}{dosing interval to be used}

\item{ndose}{number of doses to be used}

\item{times}{In case tau and ndose cannot be used (unequal dosing), this parameter can be
used to set times of dosing (e.g c(0,24,168))}
}
\value{
a data frame that can be used as an event dataset within lsoda
}
\description{
This function creates a data.frame that can be used for events of the lsoda function
}
\examples{

 dose_func(8,100,tau=48,ndose=5,tinf=2)
}
\author{
Richard Hooijmaijers
}
