% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inits.r
\name{get_inits}
\alias{get_inits}
\title{get the initial states for the differential equations}
\usage{
get_inits(lstblock)
}
\arguments{
\item{lstblock}{list with each item being a separate strucured dollor block, usually obtain from \code{\link{nmlistblock}}}
}
\value{
a named vector with the state values
}
\description{
This function will extracts the initial states for the differential equations from a NONMEM model
}
\examples{

mod    <- system.file("example_models","PK.1CMT.ORAL.mod", package = "amp.sim")
mdll   <- get_nmblock(mod,block=c("PK","DES"))
mdlls  <- nmlistblock(mdll)
get_inits(mdlls)

}
\author{
Richard Hooijmaijers
}
