% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmlistblock.r
\name{nmlistblock}
\alias{nmlistblock}
\title{convert a block to a list}
\usage{
nmlistblock(dollmodel)
}
\arguments{
\item{dollmodel}{list with each item being a separate dollar block, usually obtain from \code{\link{get_nmblock}}}
}
\value{
a list with the structured code
}
\description{
This function will convert a NONMEM block to a list including the type and separate formula parts
}
\examples{

\dontrun{nmlistblock(dollmodel)}
}
\author{
Richard Hooijmaijers
}
