% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_delete.r
\name{par_delete}
\alias{par_delete}
\title{Delete parenthesis including a numeric within a vector}
\usage{
par_delete(vect, excfun = TRUE)
}
\arguments{
\item{vect}{the vector that should be scanned for parenthesis}

\item{excfun}{exclude functions from deleting parenthesis (taken from the globally defined function list in the package)}
}
\value{
a vector with the stripped parenthesis
}
\description{
This function deletes all parenthesis from a character vector that has a numeric value inside.
This is applicable as THETA(n), ETA(n), DADT(n) is not always allowed in R coding
}
\examples{

  par_delete(c("LOG(1)","ETA(1)","EXP(2)/ETA(3)+THETA(4)"))
}
\author{
Richard Hooijmaijers
}
