% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings2df.r
\name{settings2df}
\alias{settings2df}
\title{Add settings to dataframe}
\usage{
settings2df(savedsims, leaveout = c("go", "updOpts", "sett", "refr", "tabsel"))
}
\arguments{
\item{savedsims}{reactiveValue that contains the saved simulations}

\item{leaveout}{character vector with the the elements that should be left out the dataframe}
}
\value{
a dataframe with settings
}
\description{
This function adds settings or input elements to a dataframe for displaying in app
}
\examples{

if(requireNamespace("tidyr")){
  sim1 <- list(THETA1  = 0.5, THETA2 = 1,
               alllabs =c("THETA1\%=\%DUMMY1","THETA2\%=\%DUMMY2"))
  sim2 <- list(THETA1  = 0.9, THETA2 = 1.5)
  settings2df(list(sim1 = sim1, sim2 = sim2))
} 

}
\author{
Richard Hooijmaijers
}
