% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings2df.r
\name{settings2df}
\alias{settings2df}
\title{Add settings to dataframe}
\usage{
settings2df(savedsims, leaveout = c("go", "updOpts", "sett", "refr", "tabsel"))
}
\arguments{
\item{savedsims}{reactiveValue that contains the saved simulations}

\item{leaveout}{character vector with the the elements that should be lefted out the dataframe}
}
\value{
a dataframe with settings
}
\description{
This function adds settings or input elements to a dataframe for displaying in app
}
\examples{

\dontrun{settings2df(savedsims=savedsims)}
}
\author{
Richard Hooijmaijers
}
