% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sim.r
\name{split_sim}
\alias{split_sim}
\title{Split a large NONMEM simulation in chunks}
\usage{
split_sim(data, model, locout, splitby = "ID", numout = 4)
}
\arguments{
\item{data}{character string for input dataset or a dataframe with the simulation data (see details)}

\item{model}{character string for the simulation model}

\item{locout}{character string for the location of the split input dataset and models}

\item{splitby}{character string with the variable in the data to be split on (if this is not ID it could lead to unexpected results)}

\item{numout}{the number of 'equal length' outputs to be created.}
}
\value{
split CSV and models files are written to disk
}
\description{
This function splits a large simulation dataset and accompanying model in chunks
to prevent memory problems within the simulation or to enable simulating over multiple cores
}
\details{
In general the data can be a character string defining the input dataset (as csv file!) or a dataframe
it is proposed to use a dataframe as in many cases the splitting should take place for a large problem. In case
a dataframe is used (e.g. from loading an Rdata object)  it is not necessary to import a huge csv file.
}
\examples{

\dontrun{
  split_sim(sim.input,"run1.mod",".")
}
}
\author{
Richard Hooijmaijers
}
