% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmpl_model.r
\name{tmpl_model}
\alias{tmpl_model}
\title{Get coding for template models}
\usage{
tmpl_model(tmpl)
}
\arguments{
\item{tmpl}{string with template}
}
\value{
model syntax is returned either to console or script (within Rstudio)
}
\description{
This function returns the code for some template models
}
\details{
There are templates available for 1-2 CMT PK models for IV/bolus/oral administration
as both analytical as closed form and parameterized with constants or as CL/V.
To see which templates are available in the package
run the function without arguments
}
\examples{

 tmpl_model("des1CMToralK.tmp")
}
\author{
Richard Hooijmaijers
}
