% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmpl_model.r
\name{tmpl_model}
\alias{tmpl_model}
\title{Get coding for template models}
\usage{
tmpl_model(tmpl, ret = "console")
}
\arguments{
\item{tmpl}{character with template}

\item{ret}{character indicating how the result should be returned. Currently "console", "string" and "script" are accepted}
}
\value{
model syntax is returned either to console or script (within Rstudio) or a character string
}
\description{
This function returns the code for some template models
}
\details{
There are templates available for 1-2 CMT PK models for IV/bolus/oral administration
as both analytical as closed form and parameterized with constants or as CL/V.
To see which templates are available in the package
run the function without arguments
}
\examples{

tmpl_model()
tmpl_model("ana1CMTbolusK.tmp")

}
\author{
Richard Hooijmaijers
}
